RAG_LLM/requirements.txt
flashrank
ragas
datasets
streamlit
python-dotenv
pandas
numpy>=1.26.0,<2.0.0
tqdm
pypdf
langchain
langchain-community
langchain-core
langchain-openai
langchain-upstage
langchain-chroma
chromadb
tiktoken
rank_bm25
python-dotenv
openpyxl
